hasher-matcher-actioner/demo/api/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.0
# HMA dependencies will be installed from the parent directory 